refactor: move bigint utils to `src/optimizer/util.ts`
#1421
