+++ src/ast/util.ts
@@ -191,31 +191,3 @@ export function checkIsName(ast: A.AstExpression): boolean {
 export function checkIsBoolean(ast: A.AstExpression, b: boolean): boolean {
     return ast.kind === "boolean" ? ast.value == b : false;
 }
-
-// bigint arithmetic
-
-// precondition: the divisor is not zero
-// rounds the division result towards negative infinity
-export function divFloor(a: bigint, b: bigint): bigint {
-    const almostSameSign = a > 0n === b > 0n;
-    if (almostSameSign) {
-        return a / b;
-    }
-    return a / b + (a % b === 0n ? 0n : -1n);
-}
-
-export function abs(a: bigint): bigint {
-    return a < 0n ? -a : a;
-}
-
-export function sign(a: bigint): bigint {
-    if (a === 0n) return 0n;
-    else return a < 0n ? -1n : 1n;
-}
-
-// precondition: the divisor is not zero
-// rounds the result towards negative infinity
-// Uses the fact that a / b * b + a % b == a, for all b != 0.
-export function modFloor(a: bigint, b: bigint): bigint {
-    return a - divFloor(a, b) * b;
-}
